Finish CRUD for decks

+++ app/api/deck_routes.py
@@ -14,7 +14,48 @@ def get_all_decks():
     
     return all_decks
 
-### create a deck for a class that they made
+### Edit a deck name for a class that they own
+@deck_routes.route('/<int:id>', methods = ["PUT"])
+def edit_deck_name(id):
+    if current_user.is_authenticated:
+        deck_obj = Deck.query.get(id)
+        if not deck_obj:
+            return "Deck does not exist (yet)", 404
+        
+        data = request.json
+        
+        # User must own the class that owns the deck
+        class_id = deck_obj.class_id
+        class_obj = Class.query.get(class_id)
+         
+        if class_obj.user_id != current_user.id:
+            return "User does not own class", 401
+        
+        # block user from being allowed to assign decks to classes they don't own
+        id = current_user.id
+        owned_classes = Class.query.filter(Class.user_id == id).all()
+        owned_class_ids = [class_var.user_id for class_var in owned_classes]
+        if data['class_id'] not in owned_class_ids:
+            return "User cannot assign deck to class that they don't own"
+        
+        print(owned_class_ids, "<--------------------------------")
+        
+        # change name of deck and class id
+        deck_obj.name = data['name']
+        deck_obj.class_id = data['class_id']
+                
+        db.session.add(deck_obj)
+        db.session.commit()
+        
+        result = {
+            "id": deck_obj.id,
+            "name": deck_obj.name,
+            "class_id": deck_obj.class_id
+        }
+    
+    return result
+
+### Create a deck for a class that they made
 @deck_routes.route('/', methods = ["POST"])
 def create_decks():
     if current_user.is_authenticated:
@@ -50,15 +91,14 @@ def delete_class_by_id(id):
         deck_obj = Deck.query.get(id)
         
         if not deck_obj: 
-            return "Deck not found", 401 # 401 is unathorized
+            return "Deck not found", 404 # 401 is unathorized
         
         # Make sure that the user owns the deck in question
-        # class_id = deck_obj.class_id
-        # class_obj = Class.query.get(class_id)
+        class_id = deck_obj.class_id
+        class_obj = Class.query.get(class_id)
         
-        # print(class_obj.user_id, "<------------------------- user id")
-        # if(current_user.id != class_obj.user_id):
-        #     return "User does not own class which owns deck", 401
+        if(current_user.id != class_obj.user_id):
+            return "User does not own class which owns deck", 401
                    
         try: 
             db.session.delete(deck_obj)
